Update vectorized reader to directly return ColumnarBatch instead of a row by row RDD.
This is a hack to get it to work using RDD as the main interface. This patch does this
by changing the new ParquetRecordReader to return Object, which can either be an UnsafeRow
or a ColumnarBatch. If vectorized is enabled, it always returns a ColumnarBatch. Currently,
the *only* consumer of ColumnarBatch is the codegen'd ExistingRDD which knows how to deal
with this.

Perf:
TPCDS:                             Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)
---------------------------------------------------------------------------------
q55 (before)                           10488 / 11086         11.0          91.0
q55                                      6908 / 7388         16.7          60.0